scripts: Add "audit" logs
This commit adds support to what must-gather defines as "audit"
logs, and this, confusely, doesn't mean the logs that are part
of `/var/log/audit`, but rather a collection of logs that either
must-gather or OCP defines as "audit" essential logs.

The file used here is heavily based on what we have as part of
the must-gather official repo, but instead of getting info for
the master nodes, it collects the info to the worker nodes.

The must-gather original file can be found at:
https://github.com/openshift/must-gather/blob/master/collection-scripts/gather_audit_logs

Fixes: https://github.com/fidencio/openshift-sandboxed-containers-must-gather/issues/1

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>